+++ core/initializer/test_execution/test_executor/test_executor.go
@@ -3,7 +3,7 @@
  * All Rights Reserved.
  */
 
-package parallelism
+package test_executor
 
 import (
 	"context"
@@ -60,130 +60,81 @@ type testResult struct {
 	executionErr error
 }
 
-/*
-Executor responsible for running a test with timeout, cleaning up after the test as needed.
- */
-type testExecutor struct {
-	// The logger to which all log statements must be sent
-	log *logrus.Logger
-
-	// The execution UUID that the test is running with
-	executionInstanceId uuid.UUID
-
-	// The Docker client to execute Docker actions with
-	dockerClient *client.Client
-
-	// The mask of the subnet that the test should run in
-	subnetMask string
-
-	// The name of the Docker image of the Kurtosis API container to run
-	kurtosisApiImageName string
-
-	// The name of the Docker image of the test suite to run
-	testSuiteImageName string
-
-	// The log level string to pass to the test suite (should be meaningful to the test suite image)
-	testSuiteLogLevel string
-
-	// Mapping of user-defined custom environment variables that will also be passed to the test suite contianer on start
-	customTestSuiteEnvVars map[string]string
-
-	// Name of the test being run
-	testName string
-
-	// The string representing the log level that the API container should run with
-	apiContainerLogLevel string
-}
 
 /*
-Creates a new test executor with the given params, ready to execute a single test. Technically all these params
-	could be passed to the runTest method, but it's much simpler to do these at a per-instance level since executors
-	don't get reused.
+Runs a single test with the given name
 
 Args:
-	log: the logger to which all logging events during test execution will be sent
 	executionInstanceId: The UUID representing an execution of the user's test suite, to which this test execution belongs
+	ctx: The Context that the test execution is happening in
+	log: the logger to which all logging events during test execution will be sent
 	dockerClient: The Docker client to use to manipulate the Docker engine
 	subnetMask: The subnet mask of the Docker network that has been spun up for this test
+	kurtosisApiImageName: The name of the Docker image that will be used to run the Kurtosis API container
 	testSuiteImageName: The name of the Docker image of the test controller that will orchestrate execution of this test
 	testSuiteLogLevel: A string representing the log level that the test controller should set for itself; this string
 		should be meaningful to the user-defined controller code
 	customTestSuiteEnvVars: A key-value mapping of custom Docker environment variables that will be passed to the
 		controller image (as a method for the user to pass their own custom params between initializer and controller)
 	testName: The name of the test the executor should execute
- */
-func newTestExecutor(
-			log *logrus.Logger,
-			executionInstanceId uuid.UUID,
-			dockerClient *client.Client,
-			subnetMask string,
-			kurtoisApiImageName string,
-			testSuiteImageName string,
-			testSuiteLogLevel string,
-			customTestControllerEnvVars map[string]string,
-			testName string,
-			apiContainerLogLevel string) *testExecutor {
-	return &testExecutor{
-		// TODO sort alphabetically
-		log:                             log,
-		executionInstanceId:             executionInstanceId,
-		dockerClient:                    dockerClient,
-		subnetMask:                      subnetMask,
-		kurtosisApiImageName:            kurtoisApiImageName,
-		testSuiteImageName:              testSuiteImageName,
-		testSuiteLogLevel:               testSuiteLogLevel,
-		customTestSuiteEnvVars:          customTestControllerEnvVars,
-		testName:                        testName,
-		apiContainerLogLevel: apiContainerLogLevel,
-	}
-}
 
-
-/*
-Runs a single test with the given name
+Returns:
+	bool: True if the test passed, false otherwise
+	error: Non-nil if an error occurred that prevented the test pass/fail status from being retrieved
 */
-func (executor testExecutor) runTest(ctx context.Context) (bool, error) {
-	uniqueTestIdentifier := fmt.Sprintf("%v-%v", executor.executionInstanceId.String(), executor.testName)
+func RunTest(
+		executionInstanceId uuid.UUID,
+		ctx context.Context,
+		log *logrus.Logger,
+		dockerClient *client.Client,
+		subnetMask string,
+		kurtosisApiImageName string,
+		apiContainerLogLevel string,
+		testSuiteImageName string,
+		testSuiteLogLevel string,
+		customTestSuiteEnvVars map[string]string,
+		testName string) (bool, error) {
+	uniqueTestIdentifier := fmt.Sprintf("%v-%v", executionInstanceId.String(), testName)
 
-	executor.log.Info("Creating Docker manager from environment settings...")
+	log.Info("Creating Docker manager from environment settings...")
 	// NOTE: at this point, all Docker commands from here forward will be bound by the Context that we pass in here - we'll
 	//  only need to cancel this context once
-	dockerManager, err := commons.NewDockerManager(executor.log, executor.dockerClient)
+	dockerManager, err := commons.NewDockerManager(log, dockerClient)
 	if err != nil {
-		return false, stacktrace.Propagate(err, "An error occurred getting the Docker manager for test %v", executor.testName)
+		return false, stacktrace.Propagate(err, "An error occurred getting the Docker manager for test %v", testName)
 	}
-	executor.log.Info("Docker manager created successfully")
+	log.Info("Docker manager created successfully")
 
 	volumeName := uniqueTestIdentifier
-	executor.log.Debugf("Creating Docker volume %v which will be shared with the test network...", volumeName)
+	log.Debugf("Creating Docker volume %v which will be shared with the test network...", volumeName)
 	if err := dockerManager.CreateVolume(ctx, volumeName); err != nil {
-		return false, stacktrace.Propagate(err, "Error creating Docker volume to share amongst test nodes for test %v", executor.testName)
+		return false, stacktrace.Propagate(err, "Error creating Docker volume to share amongst test nodes for test %v", testName)
 	}
-	executor.log.Debugf("Docker volume %v created successfully", volumeName)
+	log.Debugf("Docker volume %v created successfully", volumeName)
 
-	executor.log.Infof("Creating Docker network for test with subnet mask %v...", executor.subnetMask)
+	log.Infof("Creating Docker network for test with subnet mask %v...", subnetMask)
 	freeIpAddrTracker, err := commons.NewFreeIpAddrTracker(
-		executor.log,
-		executor.subnetMask,
+		log,
+		subnetMask,
 		map[string]bool{})
 	if err != nil {
-		return false, stacktrace.Propagate(err, "An error occurred creating the free IP address tracker for test %v", executor.testName)
+		return false, stacktrace.Propagate(err, "An error occurred creating the free IP address tracker for test %v", testName)
 	}
 	gatewayIp, err := freeIpAddrTracker.GetFreeIpAddr()
 	if err != nil {
-		return false, stacktrace.Propagate(err, "An error occurred getting a free IP for the gateway for test %v", executor.testName)
+		return false, stacktrace.Propagate(err, "An error occurred getting a free IP for the gateway for test %v", testName)
 	}
-	networkName := fmt.Sprintf("%v-%v", executor.executionInstanceId.String(), executor.testName)
-	networkId, err := dockerManager.CreateNetwork(ctx, networkName, executor.subnetMask, gatewayIp)
+	networkName := fmt.Sprintf("%v-%v", executionInstanceId.String(), testName)
+	networkId, err := dockerManager.CreateNetwork(ctx, networkName, subnetMask, gatewayIp)
 	if err != nil {
 		// TODO If the user Ctrl-C's while the CreateNetwork call is ongoing then the CreateNetwork will error saying
 		//  that the Context was cancelled as expected, but *the Docker engine will still create the networks!!! We'll
 		//  need to parse the log message for the string "context canceled" and, if found, do another search for
 		//  networks with our network name and delete them
-		return false, stacktrace.Propagate(err, "Error occurred creating Docker network %v for test %v", networkName, executor.testName)
+		return false, stacktrace.Propagate(err, "Error occurred creating Docker network %v for test %v", networkName, testName)
 	}
-	defer removeNetworkDeferredFunc(executor.log, dockerManager, networkId)
-	executor.log.Infof("Docker network %v created successfully", networkId)
+	defer removeNetworkDeferredFunc(log, dockerManager, networkId)
+	log.Infof("Docker network %v created successfully", networkId)
 
 	// TODO use hostnames rather than IPs, which makes things nicer and which we'll need for Docker swarm support
 	// We need to create the IP addresses for BOTH containers because the testsuite needs to know the IP of the API
@@ -197,7 +148,7 @@ func (executor testExecutor) runTest(ctx context.Context) (bool, error) {
 		return false, stacktrace.Propagate(err, "An error occurred getting an IP for the test suite container running the test")
 	}
 
-	executor.log.Debugf(
+	log.Debugf(
 		"Test suite container IP: %v; kurtosis API container IP: %v",
 		testRunningContainerIp.String(),
 		kurtosisApiIp.String())
@@ -211,19 +162,19 @@ func (executor testExecutor) runTest(ctx context.Context) (bool, error) {
 
 	testSuiteLogFilepath := path.Join(test_suite_mount_locations.BindMountsDirpath, testSuiteLogFilename)
 	testSuiteEnvVars, err := generateTestSuiteEnvVars(
-		executor.testName,
+		testName,
 		kurtosisApiIp.String(),
 		testSuiteLogFilepath,
-		executor.testSuiteLogLevel,
-		executor.customTestSuiteEnvVars)
+		testSuiteLogLevel,
+		customTestSuiteEnvVars)
 	if err != nil {
 		return false, stacktrace.Propagate(err, "An error occurred generating the map of test suite environment variables")
 	}
 
-	executor.log.Infof("Creating test suite container that will run the test...")
+	log.Infof("Creating test suite container that will run the test...")
 	testRunningContainerId, err := dockerManager.CreateAndStartContainer(
 		ctx,
-		executor.testSuiteImageName,
+		testSuiteImageName,
 		networkId,
 		testRunningContainerIp,
 		map[nat.Port]bool{},
@@ -238,13 +189,13 @@ func (executor testExecutor) runTest(ctx context.Context) (bool, error) {
 	if err != nil {
 		return false, stacktrace.Propagate(err, "An error occurred creating the test suite container to run the test")
 	}
-	executor.log.Info("Successfully created test suite container to run the test")
+	log.Info("Successfully created test suite container to run the test")
 
-	executor.log.Info("Creating Kurtosis API container...")
+	log.Info("Creating Kurtosis API container...")
 	kurtosisApiPort := nat.Port(fmt.Sprintf("%v/tcp", api_container_docker_consts.ContainerPort))
 	kurtosisApiContainerId, err := dockerManager.CreateAndStartContainer(
 		ctx,
-		executor.kurtosisApiImageName,
+		kurtosisApiImageName,
 		networkId,
 		kurtosisApiIp,
 		map[nat.Port]bool{
@@ -255,9 +206,9 @@ func (executor testExecutor) runTest(ctx context.Context) (bool, error) {
 			api_container_env_vars.ApiContainerIpAddrEnvVar:       kurtosisApiIp.String(),
 			api_container_env_vars.ApiLogFilepathEnvVar:           api_container_docker_consts.LogMountFilepath,
 			api_container_env_vars.GatewayIpEnvVar: gatewayIp.String(),
-			api_container_env_vars.LogLevelEnvVar: executor.apiContainerLogLevel,
+			api_container_env_vars.LogLevelEnvVar: apiContainerLogLevel,
 			api_container_env_vars.NetworkIdEnvVar: networkId,
-			api_container_env_vars.SubnetMaskEnvVar: executor.subnetMask,
+			api_container_env_vars.SubnetMaskEnvVar: subnetMask,
 			api_container_env_vars.TestSuiteContainerIdEnvVar: testRunningContainerId,
 			api_container_env_vars.TestSuiteContainerIpAddrEnvVar: testRunningContainerIp.String(),
 			api_container_env_vars.TestVolumeName: volumeName,
@@ -273,11 +224,11 @@ func (executor testExecutor) runTest(ctx context.Context) (bool, error) {
 	if err != nil {
 		return false, stacktrace.Propagate(err, "An error occurred creating the Kurtosis API container")
 	}
-	executor.log.Infof("Successfully created Kurtosis API container")
+	log.Infof("Successfully created Kurtosis API container")
 
 	// TODO add a timeout waiting for Kurtosis API container to stop???
 	// The Kurtosis API will be our indication of whether the test suite container stopped within the timeout or not
-	executor.log.Info("Waiting for Kurtosis API container to exit...")
+	log.Info("Waiting for Kurtosis API container to exit...")
 	kurtosisApiExitCode, err := dockerManager.WaitForExit(
 		context.Background(),
 		kurtosisApiContainerId)
@@ -292,7 +243,7 @@ func (executor testExecutor) runTest(ctx context.Context) (bool, error) {
 		// TODO switch to printing container STDOUT/STDERR, since a container failure that happens before the
 		//  Docker container launches (e.g. a variable not set) won't get logged here and thus won't show up in the output
 		// TODO this is in a really crappy spot; move it
-		banner_printer.PrintContainerLogsWithBanners(executor.log, testRunningContainerDescription, containerLogFp.Name())
+		banner_printer.PrintContainerLogsWithBanners(log, testRunningContainerDescription, containerLogFp.Name())
 	case exit_codes.OutOfOrderTestStatusExitCode:
 		testStatusRetrievalError = stacktrace.NewError("The Kurtosis API container received an out-of-order " +
 			"test execution status update; this is a Kurtosis code bug")
@@ -305,7 +256,7 @@ func (executor testExecutor) runTest(ctx context.Context) (bool, error) {
 		// TODO switch to printing container STDOUT/STDERR, since a container failure that happens before the
 		//  Docker container launches (e.g. a variable not set) won't get logged here and thus won't show up in the output
 		// TODO this is in a really crappy spot; move it
-		banner_printer.PrintContainerLogsWithBanners(executor.log, testRunningContainerDescription, containerLogFp.Name())
+		banner_printer.PrintContainerLogsWithBanners(log, testRunningContainerDescription, containerLogFp.Name())
 	case exit_codes.ShutdownSignalExitCode:
 		testStatusRetrievalError = stacktrace.NewError("The Kurtosis API container exited due to receiving " +
 			"a shutdown signal; if this is not expected, it's a Kurtosis bug")
@@ -315,10 +266,10 @@ func (executor testExecutor) runTest(ctx context.Context) (bool, error) {
 			"for the initializer", kurtosisApiExitCode)
 	}
 	if testStatusRetrievalError != nil {
-		executor.log.Error("An error occurred that prevented retrieval of the test completion status")
+		log.Error("An error occurred that prevented retrieval of the test completion status")
 		return false, testStatusRetrievalError
 	}
-	executor.log.Info("The test suite container running the test exited within the hard test timeout")
+	log.Info("The test suite container running the test exited within the hard test timeout")
 
 	// The test suite container will already have stopped, so now we get the exit code
 	testSuiteExitCode, err := dockerManager.WaitForExit(