+++ core/initializer/test_execution/test_executor_parallelizer/capturing_log_writer.go
@@ -3,7 +3,7 @@
  * All Rights Reserved.
  */
 
-package parallelism
+package test_executor_parallelizer
 
 import (
 	"runtime"