+++ core/initializer/test_suite_runner/test_suite_runner.go
@@ -10,7 +10,7 @@ import (
 	"fmt"
 	"github.com/docker/distribution/uuid"
 	"github.com/docker/docker/client"
-	"github.com/kurtosis-tech/kurtosis/initializer/parallelism"
+	"github.com/kurtosis-tech/kurtosis/initializer/test_execution/test_executor_parallelizer"
 	"github.com/kurtosis-tech/kurtosis/initializer/test_suite_metadata_acquirer"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
@@ -28,121 +28,68 @@ const (
 )
 
 /*
-An executor to run one or more tests from a given test suite
- */
-type TestSuiteRunner struct {
-	// TODO sort these properties alphabetically
-
-	// Docker client to use for interacting with the Docker engine
-	dockerClient *client.Client
-
-	// The name of image of test suite
-	testSuiteImage               string
-
-	// The name of the Kurtosis API Docker image
-	kurtosisApiImage string
-
-	// Key-value mapping that will be passed as-is to the test suite container on startup in the form of Docker
-	// 	environment variables
-	customTestSuiteEnvVars map[string]string
-
-	// A string, meaningful only to the test controller, that represents the log level that the controller container should
-	//	run with
-	testSuiteLogLevel string
-
-	// The log level the Kurtosis API container should use
-	apiContainerLogLevel string
-
-	// Metadata about the test suite - names of tests inside, network width bits, etc.
-	testSuiteMetadata test_suite_metadata_acquirer.TestSuiteMetadata
-}
-
-/*
-Creates a new TestSuiteRunner with the given parameters.
-
-Args:
-	client: Docker client to use when interacting with the Docker engine
-	testSuite: The test suite containing all the user's registered tests
-	testControllerImageName: The name of the Docker image of the test controller that will orchestrate test execution
-	testSuiteLogLevel: The string representing the loglevel of the test suite (the initializer won't be able
-		to parse this, so this should be meaningful to the test suite image)
-	networkWidthBits: Each test will get a Docker network with a number of available IP addresses = 2^network_width_bits.
-		This parameter should be set high enough so that each test can fit all the services they want.
-	apiContainerLogLevel: The log level the Kurtosis API container should use
-	testSuiteMetadata: Metadata about the test suite - e.g. name of tests, network width bits, etc.
- */
-func NewTestSuiteRunner(
-			// TODO sort these alphabetically
-			dockerClient *client.Client,
-			testSuiteImage string,
-			kurtosisApiImage string,
-			testSuiteLogLevel string,
-			customTestSuiteEnvVars map[string]string,
-			apiContainerLogLevel string,
-			testSuiteMetadata test_suite_metadata_acquirer.TestSuiteMetadata) *TestSuiteRunner {
-	return &TestSuiteRunner{
-		dockerClient:           dockerClient,
-		testSuiteImage:         testSuiteImage,
-		kurtosisApiImage:       kurtosisApiImage,
-		testSuiteLogLevel:      testSuiteLogLevel,
-		customTestSuiteEnvVars: customTestSuiteEnvVars,
-		apiContainerLogLevel:   apiContainerLogLevel,
-		testSuiteMetadata: testSuiteMetadata,
-	}
-}
-
-/*
 Runs the tests with the given names and prints the results to STDOUT. If no tests are specifically defined, all tests are run.
 
 Args:
+	dockerClient: Docker client to use when interacting with the Docker engine
+	testSuiteMetadata: Metadata about the test suite - e.g. name of tests, network width bits, etc.
 	testNamesToRun: A "set" of test names to run
 	testParallelism: How many tests to run in parallel
+	kurtosisApiImage: The name of the Docker image that will be used to launch the Kurtosis API container
+	apiContainerLogLevel: The log level the Kurtosis API container should use
+	testSuiteImage: The Docker image that will be used to launch the test suite container
+	testSuiteLogLevel: The string representing the loglevel of the test suite (the initializer won't be able
+		to parse this, so this should be meaningful to the test suite image)
+	customTestSuiteEnvVars: Key-value mapping of custom Docker environment variables that will be passed as-is to
+		the test suite container
 
 Returns:
 	allTestsPassed: True if all tests passed, false otherwise
 	executionErr: An error that will be non-nil if an error occurred that prevented the test from running and/or the result
 		being retrieved. If this is non-nil, the allTestsPassed value is undefined!
  */
-func (runner TestSuiteRunner) RunTests(testNamesToRun map[string]bool, testParallelism uint) (allTestsPassed bool, executionErr error) {
+func RunTests(
+		dockerClient *client.Client,
+		testSuiteMetadata test_suite_metadata_acquirer.TestSuiteMetadata,
+		testNamesToRun map[string]bool,
+		testParallelism uint,
+		kurtosisApiImage string,
+		apiContainerLogLevel string,
+		testSuiteImage string,
+		testSuiteLogLevel string,
+		customTestSuiteEnvVars map[string]string) (allTestsPassed bool, executionErr error) {
 	// If the user doesn't specify any test names to run, do all of them
 	if len(testNamesToRun) == 0 {
 		testNamesToRun = map[string]bool{}
-		for testName := range runner.testSuiteMetadata.TestNames {
+		for testName := range testSuiteMetadata.TestNames {
 			testNamesToRun[testName] = true
 		}
 	}
 
 	// Validate all the requested tests exist
 	for testName := range testNamesToRun {
-		if _, found := runner.testSuiteMetadata.TestNames[testName]; !found {
+		if _, found := testSuiteMetadata.TestNames[testName]; !found {
 			return false, stacktrace.NewError("No test registered with name '%v'", testName)
 		}
 	}
 
 	executionInstanceId := uuid.Generate()
-	testParams, err := buildTestParams(executionInstanceId, testNamesToRun, runner.testSuiteMetadata.NetworkWidthBits)
+	testParams, err := buildTestParams(executionInstanceId, testNamesToRun, testSuiteMetadata.NetworkWidthBits)
 	if err != nil {
 		return false, stacktrace.Propagate(err, "An error occurred building the test params map")
 	}
 
-	// Initialize a Docker client
-	dockerClient, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
-	if err != nil {
-		return false, stacktrace.Propagate(err,"Failed to initialize Docker client from environment.")
-	}
-
-	testExecutor := parallelism.NewTestExecutorParallelizer(
+	logrus.Infof("Running %v tests with execution ID %v...", len(testNamesToRun), executionInstanceId.String())
+	allTestsPassed = test_executor_parallelizer.RunInParallelAndPrintResults(
 		executionInstanceId,
 		dockerClient,
-		runner.kurtosisApiImage,
-		runner.testSuiteImage,
-		runner.testSuiteLogLevel,
-		runner.customTestSuiteEnvVars,
 		testParallelism,
-		runner.apiContainerLogLevel)
-
-	logrus.Infof("Running %v tests with execution ID %v...", len(testNamesToRun), executionInstanceId.String())
-	allTestsPassed = testExecutor.RunInParallelAndPrintResults(testParams)
+		testParams,
+		kurtosisApiImage,
+		apiContainerLogLevel,
+		testSuiteImage,
+		testSuiteLogLevel,
+		customTestSuiteEnvVars)
 	return allTestsPassed, nil
 }
 
@@ -152,7 +99,7 @@ Helper function to build, from the set of tests to run, the map of test params t
 Args:
 	testsToRun: A "set" of test names to run in parallel
  */
-func buildTestParams(executionInstanceId uuid.UUID, testNamesToRun map[string]bool, networkWidthBits uint32) (map[string]parallelism.ParallelTestParams, error) {
+func buildTestParams(executionInstanceId uuid.UUID, testNamesToRun map[string]bool, networkWidthBits uint32) (map[string]test_executor_parallelizer.ParallelTestParams, error) {
 	subnetMaskBits := BITS_IN_IP4_ADDR - networkWidthBits
 
 	subnetStartIp := net.ParseIP(SUBNET_START_ADDR)
@@ -171,7 +118,7 @@ func buildTestParams(executionInstanceId uuid.UUID, testNamesToRun map[string]bo
 	}
 
 	testIndex := 0
-	testParams := make(map[string]parallelism.ParallelTestParams)
+	testParams := make(map[string]test_executor_parallelizer.ParallelTestParams)
 	for testName, _ := range testNamesToRun {
 		// Pick the next free available subnet IP, considering all the tests we've started previously
 		subnetIpInt := subnetStartIpInt + uint32(testIndex) * uint32(math.Pow(2, float64(networkWidthBits)))
@@ -179,7 +126,7 @@ func buildTestParams(executionInstanceId uuid.UUID, testNamesToRun map[string]bo
 		binary.BigEndian.PutUint32(subnetIp, subnetIpInt)
 		subnetCidrStr := fmt.Sprintf("%v/%v", subnetIp.String(), subnetMaskBits)
 
-		testParams[testName] = *parallelism.NewParallelTestParams(testName, subnetCidrStr, executionInstanceId)
+		testParams[testName] = *test_executor_parallelizer.NewParallelTestParams(testName, subnetCidrStr, executionInstanceId)
 		testIndex++
 	}
 	return testParams, nil