+++ core/CHANGELOG.md
@@ -20,6 +20,7 @@
 * Remove TODO in run.sh about parameterizing binary name
 * Allow configurable, custom Docker environment variables that will be passed as-is to the test suite
 * Added `--list` arg to print test names in test suite
+* Kill unnecessary `TestSuiteRunner`, `TestExecutorParallelizer`, and `TestExecutor` structs
 
 # 0.9.0
 * Change ConfigurationID to be a string